[IMP] accounting/electronic_invoicing: adapt to new Send & Print
From saas-16.2, the choice of the format can be made from the partner
form view (for the formats present in `account_edi_ubl_cii` only). And
the attachment are generated during the Send & Print fow.

PR to select the EDI format based on the partner:
https://github.com/odoo/odoo/pull/115934

PR refactoring the Send & Print flow:
https://github.com/odoo/odoo/pull/111857